+++ api/internal/handler/routes.go
@@ -189,6 +189,16 @@ func RegisterHandlers(server *rest.Server, serverCtx *svc.ServiceContext) {
 					Path:    "/user/logout",
 					Handler: user.LogoutHandler(serverCtx),
 				},
+				{
+					Method:  http.MethodGet,
+					Path:    "/user/refresh_token",
+					Handler: user.RefreshTokenHandler(serverCtx),
+				},
+				{
+					Method:  http.MethodGet,
+					Path:    "/user/access_token",
+					Handler: user.AccessTokenHandler(serverCtx),
+				},
 			}...,
 		),
 		rest.WithJwt(serverCtx.Config.Auth.AccessSecret),