+++ api/internal/handler/user/refresh_token_handler.go
@@ -1,0 +1,32 @@
+package user
+
+import (
+	"net/http"
+
+	"github.com/zeromicro/go-zero/rest/httpx"
+
+	"github.com/suyuan32/simple-admin-core/api/internal/logic/user"
+	"github.com/suyuan32/simple-admin-core/api/internal/svc"
+)
+
+// swagger:route get /user/refresh_token user RefreshToken
+//
+// Refresh token | 获取刷新 token
+//
+// Refresh token | 获取刷新 token
+//
+// Responses:
+//  200: RefreshTokenResp
+
+func RefreshTokenHandler(svcCtx *svc.ServiceContext) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		l := user.NewRefreshTokenLogic(r.Context(), svcCtx)
+		resp, err := l.RefreshToken()
+		if err != nil {
+			err = svcCtx.Trans.TransError(r.Context(), err)
+			httpx.ErrorCtx(r.Context(), w, err)
+		} else {
+			httpx.OkJsonCtx(r.Context(), w, resp)
+		}
+	}
+}