+++ api/internal/logic/menu/update_menu_logic.go
@@ -36,6 +36,7 @@ func (l *UpdateMenuLogic) UpdateMenu(req *types.MenuPlainInfo) (resp *types.Base
 		Sort:        req.Sort,
 		Disabled:    req.Disabled,
 		ServiceName: req.ServiceName,
+		Permission:  req.Permission,
 		Meta: &core.Meta{
 			Title:              req.Title,
 			Icon:               req.Icon,