+++ rpc/ent/schema/menu.go
@@ -18,69 +18,51 @@ type Menu struct {
 func (Menu) Fields() []ent.Field {
 	return []ent.Field{
 		field.Uint64("parent_id").Default(100000).Optional().
-			Comment("Parent menu ID | 父菜单ID").
-			Annotations(entsql.WithComments(true)),
+			Comment("Parent menu ID | 父菜单ID"),
 		field.Uint32("menu_level").
-			Comment("Menu level | 菜单层级").
-			Annotations(entsql.WithComments(true)),
+			Comment("Menu level | 菜单层级"),
 		field.Uint32("menu_type").
-			Comment("Menu type | 菜单类型 （菜单或目录）0 目录 1 菜单").
-			Annotations(entsql.WithComments(true)),
+			Comment("Menu type | 菜单类型 （菜单或目录）0 目录 1 菜单"),
 		field.String("path").Optional().Default("").
-			Comment("Index path | 菜单路由路径").
-			Annotations(entsql.WithComments(true)),
+			Comment("Index path | 菜单路由路径"),
 		field.String("name").
-			Comment("Index name | 菜单名称").
-			Annotations(entsql.WithComments(true)),
+			Comment("Index name | 菜单名称"),
 		field.String("redirect").Optional().Default("").
-			Comment("Redirect path | 跳转路径 （外链）").
-			Annotations(entsql.WithComments(true)),
+			Comment("Redirect path | 跳转路径 （外链）"),
 		field.String("component").Optional().Default("").
-			Comment("The path of vue file | 组件路径").
-			Annotations(entsql.WithComments(true)),
+			Comment("The path of vue file | 组件路径"),
 		field.Bool("disabled").Optional().Default(false).
-			Comment("Disable status | 是否停用").
-			Annotations(entsql.WithComments(true)),
+			Comment("Disable status | 是否停用"),
 		field.String("service_name").Optional().Default("Other").
-			Comment("Service Name | 服务名称").
-			Annotations(entsql.WithComments(true)),
+			Comment("Service Name | 服务名称"),
+		field.String("permission").
+			Optional().
+			Comment("Permission symbol | 权限标识"),
 		// meta
 		field.String("title").
-			Comment("Menu name | 菜单显示标题").
-			Annotations(entsql.WithComments(true)),
+			Comment("Menu name | 菜单显示标题"),
 		field.String("icon").
-			Comment("Menu icon | 菜单图标").
-			Annotations(entsql.WithComments(true)),
+			Comment("Menu icon | 菜单图标"),
 		field.Bool("hide_menu").Optional().Default(false).
-			Comment("Hide menu | 是否隐藏菜单").
-			Annotations(entsql.WithComments(true)),
+			Comment("Hide menu | 是否隐藏菜单"),
 		field.Bool("hide_breadcrumb").Optional().Default(false).
-			Comment("Hide the breadcrumb | 隐藏面包屑").
-			Annotations(entsql.WithComments(true)),
+			Comment("Hide the breadcrumb | 隐藏面包屑"),
 		field.Bool("ignore_keep_alive").Optional().Default(false).
-			Comment("Do not keep alive the tab | 取消页面缓存").
-			Annotations(entsql.WithComments(true)),
+			Comment("Do not keep alive the tab | 取消页面缓存"),
 		field.Bool("hide_tab").Optional().Default(false).
-			Comment("Hide the tab header | 隐藏页头").
-			Annotations(entsql.WithComments(true)),
+			Comment("Hide the tab header | 隐藏页头"),
 		field.String("frame_src").Optional().Default("").
-			Comment("Show iframe | 内嵌 iframe").
-			Annotations(entsql.WithComments(true)),
+			Comment("Show iframe | 内嵌 iframe"),
 		field.Bool("carry_param").Optional().Default(false).
-			Comment("The route carries parameters or not | 携带参数").
-			Annotations(entsql.WithComments(true)),
+			Comment("The route carries parameters or not | 携带参数"),
 		field.Bool("hide_children_in_menu").Optional().Default(false).
-			Comment("Hide children menu or not | 隐藏所有子菜单").
-			Annotations(entsql.WithComments(true)),
+			Comment("Hide children menu or not | 隐藏所有子菜单"),
 		field.Bool("affix").Optional().Default(false).
-			Comment("Affix tab | Tab 固定").
-			Annotations(entsql.WithComments(true)),
+			Comment("Affix tab | Tab 固定"),
 		field.Uint32("dynamic_level").Optional().Default(20).
-			Comment("The maximum number of pages the router can open | 能打开的子TAB数").
-			Annotations(entsql.WithComments(true)),
+			Comment("The maximum number of pages the router can open | 能打开的子TAB数"),
 		field.String("real_path").Optional().Default("").
-			Comment("The real path of the route without dynamic part | 菜单路由不包含参数部分").
-			Annotations(entsql.WithComments(true)),
+			Comment("The real path of the route without dynamic part | 菜单路由不包含参数部分"),
 	}
 }
 
@@ -107,6 +89,7 @@ func (Menu) Edges() []ent.Edge {
 
 func (Menu) Annotations() []schema.Annotation {
 	return []schema.Annotation{
+		entsql.WithComments(true),
 		entsql.Annotation{Table: "sys_menus"},
 	}
 }